backend/requirements.txt
Django>=4.0
djangorestframework>=3.13
python-dateutil>=2.8.2
